sink(ticdc): move code
